+++ src/Telegram.Bot/Types/ReplyMarkups/Buttons/InlineKeyboardCallbackButton.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
-namespace Telegram.Bot.Types.InlineKeyboardButtons
+namespace Telegram.Bot.Types.ReplyMarkups.Buttons
 {
     /// <summary>
     /// This object represents one button of an inline keyboard that sends a callback query to the bot when pressed.
@@ -10,16 +10,24 @@ namespace Telegram.Bot.Types.InlineKeyboardButtons
     public class InlineKeyboardCallbackButton : InlineKeyboardButton
     {
         /// <summary>
-        /// Optional. Data to be sent in a callback query to the bot when button is pressed
+        /// Data to be sent in a callback query to the bot when button is pressed
         /// </summary>
         [JsonProperty(Required = Required.Always)]
         public string CallbackData { get; }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="InlineKeyboardButton"/> class.
+        /// Initializes a new instance of the <see cref="InlineKeyboardCallbackButton"/> class.
         /// </summary>
-        /// <param name="text">The text.</param>
-        /// <param name="callbackData">The callback data.</param>
+        /// <param name="textAndCallbackData">The label text on the button and callback data.</param>
+        public InlineKeyboardCallbackButton(string textAndCallbackData)
+            : this(textAndCallbackData, textAndCallbackData)
+        { }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="InlineKeyboardCallbackButton"/> class.
+        /// </summary>
+        /// <param name="text">Label text on the button</param>
+        /// <param name="callbackData">The callback data</param>
         public InlineKeyboardCallbackButton(string text, string callbackData)
             : base(text)
         {