+++ src/Telegram.Bot/Types/ReplyMarkups/Buttons/InlineKeyboardSwitchInlineQueryButton.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
-namespace Telegram.Bot.Types.InlineKeyboardButtons
+namespace Telegram.Bot.Types.ReplyMarkups.Buttons
 {
     /// <summary>
     /// This object represents one button of an inline keyboard that asks the user to switch to a chat
@@ -11,7 +11,7 @@ namespace Telegram.Bot.Types.InlineKeyboardButtons
     public class InlineKeyboardSwitchInlineQueryButton : InlineKeyboardButton
     {
         /// <summary>
-        /// Optional. If set, pressing the button will prompt the user to select one of their chats, open that chat and insert the bot's username and the specified <see cref="InlineQuery"/> in the input field. Can be empty, in which case just the bot's username will be inserted.
+        /// If set, pressing the button will prompt the user to select one of their chats, open that chat and insert the bot's username and the specified <see cref="InlineQuery"/> in the input field. Can be empty, in which case just the bot's username will be inserted.
         /// </summary>
         /// <remarks>
         /// Note: This offers an easy way for users to start using your bot in inline mode when they are currently in a private chat with it. Especially useful when combined with switchPm[...] parameters (see <see cref="TelegramBotClient.AnswerInlineQueryAsync"/>)  – in this case the user will be automatically returned to the chat they switched from, skipping the chat selection screen.
@@ -19,15 +19,15 @@ namespace Telegram.Bot.Types.InlineKeyboardButtons
         [JsonProperty(Required = Required.Always)]
         public string SwitchInlineQuery { get; }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="InlineKeyboardButton"/> class.
-        /// </summary>
-        /// <param name="text">Text of the button</param>
-        /// <param name="switchInlineQuery">Pressing the button will prompt the user to select one of their chats, open that chat and insert the bot‘s username and the specified inline query in the input field. Can be empty, in which case just the bot's username will be inserted.</param>
-        public InlineKeyboardSwitchInlineQueryButton(string text, string switchInlineQuery = "")
+        ///  <summary>
+        /// Initializes a new instance of the <see cref="InlineKeyboardSwitchInlineQueryButton"/> class
+        ///  </summary>
+        /// <param name="query">Input query that will be inserted with the bot's username. Can be empty, in which case just the bot’s username will be inserted.</param>
+        /// <param name="text"></param>
+        public InlineKeyboardSwitchInlineQueryButton(string text, string query = "")
             : base(text)
         {
-            SwitchInlineQuery = switchInlineQuery;
+            SwitchInlineQuery = query;
         }
     }
 }