+++ src/Telegram.Bot/Types/ReplyMarkups/Buttons/InlineKeyboardUrlButton.cs
@@ -1,37 +1,39 @@
-﻿using Newtonsoft.Json;
+﻿using System;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
-using System;
 
-namespace Telegram.Bot.Types.InlineKeyboardButtons
+namespace Telegram.Bot.Types.ReplyMarkups.Buttons
 {
     /// <summary>
     /// This object represents one button of an inline keyboard that opens an url when pressed.
     /// </summary>
-    [JsonObject(MemberSerialization = MemberSerialization.OptIn,
-                NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
+    [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
     public class InlineKeyboardUrlButton : InlineKeyboardButton
     {
         /// <summary>
         /// Optional. HTTP url to be opened when button is pressed
         /// </summary>
-        [JsonProperty]
-        public string Url { get; set; }
+        [JsonProperty(Required = Required.Always)]
+        public string Url { get; }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="InlineKeyboardButton"/> class.
+        /// Initializes a new instance of the <see cref="InlineKeyboardUrlButton"/> class.
         /// </summary>
         /// <param name="text">Text of the button</param>
         /// <param name="url">HTTP url to be opened when button is pressed</param>
-        public InlineKeyboardUrlButton(string text, string url) : base(text)
+        public InlineKeyboardUrlButton(string text, string url)
+            : base(text)
         {
             Url = url;
         }
+
         /// <summary>
-        /// Initializes a new instance of the <see cref="InlineKeyboardButton"/> class.
+        /// Initializes a new instance of the <see cref="InlineKeyboardUrlButton"/> class.
         /// </summary>
-        /// <param name="text">Text of the button</param>
+        /// <param name="text">Label text on the button</param>
         /// <param name="url">HTTP url to be opened when button is pressed</param>
-        public InlineKeyboardUrlButton(string text, Uri url) : base(text)
+        public InlineKeyboardUrlButton(string text, Uri url)
+            : base(text)
         {
             Url = url.AbsoluteUri;
         }